TicketTerminal API v7

+++ src/main/java/com/ticket/terminal/dto/CostResponseDto.java
@@ -15,5 +15,5 @@ import lombok.*;
 public class CostResponseDto {
 
     @JsonProperty("Cost")
-    private Integer cost;
+    private Double cost;
 }
